fix(contracts): resolve all clippy warnings and compile errors

+++ contracts/vault/src/storage.rs
@@ -868,7 +868,7 @@ pub fn get_pending_timelocked_proposals(env: &Env, offset: u64, limit: u32) -> V
     let mut skipped: u64 = 0;
 
     for i in 0..ids.len() {
-        if result.len() as u32 >= cap {
+        if result.len() >= cap {
             break;
         }
         let id = match ids.get(i) {
@@ -1977,7 +1977,7 @@ fn update_low_participation_state(
     config: &Config,
 ) -> (u32, bool) {
     let rate = compute_participation_rate(score, config.participation_rate_window);
-    if (rate as u32) < config.min_participation_rate {
+    if rate < config.min_participation_rate {
         score.consecutive_low_periods += 1;
         if score.low_participation_since_ledger.is_none() {
             score.low_participation_since_ledger = Some(env.ledger().sequence());
